% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_path_indicator.R
\name{add_path_indicator}
\alias{add_path_indicator}
\title{Add saving path to indicator}
\usage{
add_path_indicator(path = NULL, create_dir = TRUE)
}
\arguments{
\item{path}{A directory to save the indicators to. If \code{NULL} then
creates a \code{tempdir()} to save the data.}
}
\value{
Returns the updated \code{path}.
}
\description{
Checks the saving path to the indicator, and if it does not exist,
it tries to create it. If it fails, it returns a temporary directory
for the current R session.
}
