% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_interpolate.R
\name{approximate_interpolate}
\alias{approximate_interpolate}
\title{Approximate Indicator}
\usage{
approximate_interpolate(dat, maxgap = 3, approx_method = "linear")
}
\arguments{
\item{dat}{Eurostat data frame or a data frame derived from
\code{\link{create_indicator}} or \code{\link{create_tables}}}

\item{maxgap}{See code{\link[zoo]{na.locf}}. Defaults to \code{Inf}.}

\item{approx_method}{Defaults to \code{'linear'}. You can use \code{'constant'}.
See \code{\link[stat]{approxfun}}}
}
\description{
Interpolate missing values between known observations.
}
\examples{
{
example_df <- data.frame (
  geo = c(rep("AT", 4), rep("DE", 4)),
  time = rep(as.Date(c("2008-01-01", "2009-01-01",
  "2010-01-01", "2011-01-01")), 2),
  values = c( 1:4, NA_real_, NA_real_, 5,6)
 )
approximate_interpolate( example_df )
}
}
\seealso{
\code{\link[stat]{approxfun}} in the \code{stats} package.

Other approximate functions: 
\code{\link{approximate_locf}()},
\code{\link{approximate_nocb}()}
}
\concept{approximate functions}
