% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_nocb.R
\name{approximate_nocb}
\alias{approximate_nocb}
\title{Next Observation Carried Back}
\usage{
approximate_nocb(dat, maxgap = 2)
}
\arguments{
\item{dat}{Eurostat data frame}

\item{maxgap}{The maximum number of missing observations that
should be filled up with observation carry back.}
}
\description{
Next observation carry back in the values of the indicator.
}
\examples{
{
example_df <- data.frame (
  geo = c(rep("AT", 4), rep("DE", 4)),
  time = rep(as.Date(c("2008-01-01", "2009-01-01",
  "2010-01-01", "2011-01-01")), 2),
  values = c( 1:4, NA_real_, NA_real_, 5,6)
 )
approximate_nocb( example_df )
}
}
\seealso{
Other approximate functions: 
\code{\link{approximate_interpolate}()},
\code{\link{approximate_locf}()}
}
\concept{approximate functions}
