% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata_table.R
\name{create_eurostat_metadata_table}
\alias{create_eurostat_metadata_table}
\title{Create a Eurostat metadata table}
\usage{
create_eurostat_metadata_table(dat, code = "indicator")
}
\arguments{
\item{dat}{A data.frame acquired by \code{\link[eurostat]{get_eurostat}} and created
by \code{create_indicator}.}

\item{code}{A Eurostat product code, if omitted, defaults to
\code{'indicator'}.}
}
\value{
A data.frame containing the descriptive metadata and unit
information of the Eurostat product.
}
\description{
Create a Eurostat metadata table
}
\seealso{
Other Eurostat functions: 
\code{\link{create_eurostat_tables}()},
\code{\link{create_eurostat_toc}()},
\code{\link{create_missings}()},
\code{\link{process_eurostat_indicator}()}
}
\concept{Eurostat functions}
