% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_eurostat_source}
\alias{create_eurostat_source}
\title{Create Eurostat Source Description}
\usage{
create_eurostat_source(code)
}
\description{
Create Eurostat Source Description
}
\keyword{internal,}
\keyword{metadata}
