% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eurostat_tables.R
\name{create_eurostat_tables}
\alias{create_eurostat_tables}
\title{Create data tables from Eurostat with valid geo codes}
\usage{
create_eurostat_tables(
  eurostat_table,
  code = "indicator",
  min_time = as.Date("2010-01-01"),
  max_time = as.Date("2019-01-01")
)
}
\arguments{
\item{eurostat_table}{A data.frame acquired by \code{\link[eurostat]{get_eurostat}}
and created by \code{\link{create_indicator}}.}

\item{code}{A Eurostat product code, if omitted, defaults to
\code{'indicator'}.}

\item{min_time}{Defaults to \code{as.Date("2010-01-01")}, the earliest
time values that used to create explicit missings. If \code{NULL},
it will use the earliest observation present in \code{eurostat_table}.}

\item{max_time}{Defaults to \code{as.Date("2019-01-01")}, the latest
time values that used to create explicit missings. If \code{NULL},
it will use the latest observation present in \code{eurostat_table}.}
}
\value{
A data frame augmented with NUTS2016 codes, NUTS levels,
and all factor variables converted to characters.
}
\description{
Create data tables from Eurostat with valid geo codes
}
\seealso{
\code{\link[eurostat]{get_eurostat}}

Other Eurostat functions: 
\code{\link{create_eurostat_metadata_table}()},
\code{\link{create_eurostat_toc}()},
\code{\link{create_missings}()},
\code{\link{process_eurostat_indicator}()}

Other Eurostat functions: 
\code{\link{create_eurostat_metadata_table}()},
\code{\link{create_eurostat_toc}()},
\code{\link{create_missings}()},
\code{\link{process_eurostat_indicator}()}
}
\concept{Eurostat functions}
