% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_indicators.R
\name{create_indicators}
\alias{create_indicators}
\title{A temporary wrapper for creating indicators}
\usage{
create_indicators(path = NULL)
}
\arguments{
\item{path}{A directory to save the indicators to. If \code{NULL} then
creates a \code{tempdir()} to save the data.}
}
\value{
The function does not return a value, only saves the indicators
to \code{path}.
}
\description{
Download the demographic data from Eurostat.
}
\examples{
\dontrun{create_indicators()} 
}
