% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_missings.R
\name{create_missings}
\alias{create_missings}
\title{Create Missing NUTS Data}
\usage{
create_missings(
  dat,
  level = 2,
  geo_var = "geo",
  time_var = "time",
  values_var = "values",
  all_time = FALSE,
  all_geo = TRUE,
  min_time = as.Date("2010-01-01"),
  max_time = as.Date("2020-01-01")
)
}
\arguments{
\item{dat}{Eurostat data frame or a data frame derived from
\code{\link{create_indicator}} or \code{\link{create_tables}}}

\item{level}{The NUTS level where the missing values need to be
identified, defaults to \code{2}.}

\item{geo_var}{Defaults to \code{'geo'}}

\item{time_var}{Defaults to \code{'time'}}

\item{all_time}{Defaults to \code{TRUE} and creates explicit missings
for all available years.  \code{FALSE} creates only for years that are present in
\code{dat}.}

\item{all_geo}{Defaults to \code{TRUE} and creates explicit missings
for all years.  \code{FALSE} creates only for years that are present in
\code{dat}.}

\item{min_time}{Defaults to \code{as.Date("2010-01-01")}, the earliest
time values that used to create explicit missings. If \code{NULL},
it will use the earliest observation present in \code{dat}.}

\item{max_time}{Defaults to \code{as.Date("2020-01-01")}, the latest
time values that used to create explicit missings. If \code{NULL},
it will use the latest observation present in \code{dat}.}
}
\description{
Returns values on all NUTS levels, and
explicitly fills up missing values that are not present on
NUTS1 or NUTS2 or NUTS3 levels.
}
\seealso{
Other Eurostat functions: 
\code{\link{create_eurostat_metadata_table}()},
\code{\link{create_eurostat_tables}()},
\code{\link{create_eurostat_toc}()},
\code{\link{process_eurostat_indicator}()}
}
\concept{Eurostat functions}
