% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_toc.R
\name{create_toc}
\alias{create_toc}
\title{Create Table Of Contents

A wrapper around \code{\link[eurostat]{get_eurostat_toc}}}
\usage{
create_toc(codes)
}
\arguments{
\item{codes}{Eurostat product codes.}
}
\description{
Create Table Of Contents

A wrapper around \code{\link[eurostat]{get_eurostat_toc}}
}
