% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-eurostat_tables.R
\docType{data}
\name{eurostat_datasets}
\alias{eurostat_datasets}
\title{Tables to refresh: Eurostat}
\format{
A data frame with 2 rows and 3 variables:
\describe{
\item{eurostat_code}{The statistical product code in the Eurostat data warehouse}
\item{pillar}{Demo Music Observatory Pillar}
\item{topic}{Topic within pillar}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
eurostat_datasets
}
\description{
A dataset for refreshing Eurostat data
}
\keyword{datasets}
