% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat_household_consumption.R
\name{eurostat_household_consumption}
\alias{eurostat_household_consumption}
\title{Example Downloader}
\usage{
eurostat_household_consumption(path)
}
\description{
Download the demographic data from Eurostat.
}
\examples{
\dontrun{national_accounts_data()} 
}
