% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{indicator}
\alias{indicator}
\alias{is.indicator}
\alias{print.indicator}
\title{Indicator for Music Observatory}
\usage{
indicator(
  x,
  shortcode,
  description,
  date_created,
  date_earliest,
  date_latest,
  original_source = NA_character_,
  original_code = NA_character_,
  doi = NA_character_,
  doi_url = NA_character_,
  keyword1,
  keyword2,
  keyword3,
  keywords = NA_character_
)

is.indicator(x)

\method{print}{indicator}(x, ...)
}
\arguments{
\item{x}{A data.frame or tibble.}

\item{shortcode}{A short, unique, programatically usable indicator
ID.}

\item{description}{A precise character string describing the indicator
for the data catalogue of the data observatory.}

\item{date_created}{The creation day or the refreshment day of the indicator.}

\item{date_earliest}{The date of the earliest observation. Can be a numeric containing a year.}

\item{date_latest}{The date of the latest observation.}

\item{original_source}{Defaults to {NA_character_}.}

\item{original_code}{The identifier in the original source, if applicable.}

\item{doi}{The digital object identifier (obtained from a provider)}

\item{doi_url}{The URL which points to the uploaded object (in the doi provider inventory)}

\item{keyword1}{The first keywords, must be one of the pillars of the
observatory.}

\item{keyword2}{The second keyword must be a topic within a pillar.}

\item{keyword3}{A search term within the topic.}

\item{keywords}{A character vector of any optional, further keywords.}
}
\value{
A data.frame or tibble with indicator attributes.
}
\description{
The class inherits all methods from a data frame, but has
many fixed attributes.
}
\examples{
test_indicator <- indicator (
                     x <- data.frame ( 
                     geo = rep(c("NL", "BE", "LU"), 4), 
                     time = rep(c(2016:2019),3), 
                     values = runif(12, 1,100)
                     ),
   shortcode = "observatory_test_1", 
   description = "A test indicator with random numbers", 
   date_created = as.Date ( "2020-08-24"),
   date_earliest  = min (x$time, na.rm=TRUE),
   date_latest  =  max(x$time, na.rm=TRUE),
   doi = NA_character_,
   doi_url = NA_character_,
   keyword1 = "test",  keyword2 = "random",  keyword3 = "Benelux"
)

## Only the first 10 observations are printed 
str( test_indicator )
}
