% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicobservatory.R
\docType{package}
\name{musicobservatory}
\alias{musicobservatory}
\title{musicobservatory: A package to create reproducibly the Music Observatory data}
\description{
\url{https://http://music.dataobservatory.eu/}
}
\section{refresh functions}{

The refresh functions check if new data is available from a public
data resource or API, and if necessary, they refresh the data.
}

