% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography_eurostat.R
\name{national_accounts_eurostat}
\alias{national_accounts_eurostat}
\title{Download Demographic Data From Eurostat}
\usage{
national_accounts_eurostat()
}
\description{
Download the demographic data from Eurostat.
}
\examples{
\dontrun{national_accounts_data()} 
}
