% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_eurostat_indicator.R
\name{process_eurostat_indicator}
\alias{process_eurostat_indicator}
\title{Process Eurostat Indicator}
\usage{
process_eurostat_indicator(eurostat_table, code = "indicator")
}
\arguments{
\item{eurostat_table}{A data.frame acquired by
\code{\link[eurostat]{get_eurostat}}.}

\item{code}{A Eurostat product code, if omitted, defaults to
\code{'indicator'}.}
}
\value{
A data.frame with unique indicator observations made of various
metadata columns, such as {\code{unit}} or {\code{selection_criteria}}.
}
\description{
Create a table with unique observations.
}
\seealso{
Other Eurostat functions: 
\code{\link{create_eurostat_metadata_table}()},
\code{\link{create_eurostat_tables}()},
\code{\link{create_eurostat_toc}()},
\code{\link{create_missings}()}
}
\concept{Eurostat functions}
\concept{create functions}
