% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_toc_elements}
\alias{select_toc_elements}
\title{Small utils functions, not to be exported.}
\usage{
select_toc_elements(toc, code)
}
\arguments{
\item{dat}{A data.frame acquired by eurostat::get_eurostat and created
by create_indicator}

\item{values_var}{The name of the column that has the indicator values,
defaults to \code{'values'}.}
}
\description{
Small utils functions, not to be exported.
}
\keyword{internal}
